requirements.txt
django-pyodbc-azure
wfastcgi
pillow

